Simplify gulfile.js by moving the scripts to separate files per function Fixes #794
